# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Zord backend server.",
        "title": "Swagger Zord API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/roles": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Get Roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Auth",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_roles_LIST.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            }
        },
        "/roles/add_resource_permission": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Auth",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body model",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_roles_ADD_RESOURCE_PERMISSION_TO_ROLE.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_roles_ADD_RESOURCE_PERMISSION_TO_ROLE.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            }
        },
        "/roles/deleteFromUser/{role}/{userId}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Auth",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role Name",
                        "name": "role",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_roles_ADD_RESOURCE_PERMISSION_TO_ROLE.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            }
        },
        "/roles/deletePermission/{role}/{resource}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Auth",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role Name",
                        "name": "role",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Resource Name",
                        "name": "resource",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_roles_ADD_RESOURCE_PERMISSION_TO_ROLE.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            }
        },
        "/roles/setToUser": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Auth",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body model",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_roles_SET_ROLE_TO_USER.Data"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_roles_ADD_RESOURCE_PERMISSION_TO_ROLE.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            }
        },
        "/roles/{role}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Roles"
                ],
                "summary": "Roles",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Auth",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role Name",
                        "name": "role",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_roles_ADD_RESOURCE_PERMISSION_TO_ROLE.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "go-skeleton_internal_application_services.Error": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Message": {},
                "Status": {
                    "type": "integer"
                }
            }
        },
...
        "go-skeleton_internal_application_services_roles_ADD_RESOURCE_PERMISSION_TO_ROLE.Data": {
            "type": "object",
            "properties": {
                "Actions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "Resource": {
                    "type": "string"
                },
                "Role": {
                    "type": "string"
                }
            }
        },
        "go-skeleton_internal_application_services_roles_ADD_RESOURCE_PERMISSION_TO_ROLE.Response": {
            "type": "object",
            "properties": {
                "Result": {
                    "type": "string"
                }
            }
        },
        "go-skeleton_internal_application_services_roles_LIST.Response": {
            "type": "object",
            "properties": {
                "Roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "go-skeleton_internal_application_services_roles_SET_ROLE_TO_USER.Data": {
            "type": "object",
            "properties": {
                "RolesName": {
                    "type": "string"
                },
                "UserId": {
                    "type": "string"
                }
            }
        },
...
    }
}